embedder_traits

Module input_events

Structs

GamepadIndex

Index of gamepad in list of system’s connected gamepads.

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad.

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad.

InputEventAndId

InputEventId

An opaque id for an InputEvent.

InputEventOutcome

InputEventResult

Flags representing the state of an InputEvent after Servo has handled it.

KeyboardEvent

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

TouchEvent

TouchId

An opaque identifier for a touch point.

WheelDelta

The wheel event deltas for every direction.

WheelEvent

Enums

EditingActionEvent

An editing action that should be performed on a WebView.

GamepadEvent

The types of Gamepad event.

GamepadUpdateType

The type of Gamepad input being updated.

ImeEvent

The types of an input method event.

InputEvent

An input event that is sent from the embedder to Servo.

MouseButton

The types of mouse buttons.

MouseButtonAction

The types of mouse events.

TouchEventType

The type of input represented by a multi-touch event.

TouchPointerType

Distinguishes the kind of physical input that produced a TouchEvent. Servo routes both pen and finger-touch input through the touch event path, but the originating subtype determines the pointerType reported to script.

WheelMode

Unit of a WheelDelta.

Statics

INPUT_EVENT_ID 🔒