Crate embedder_traits

Modules

resources

Structs

EmbedderProxy
Sends messages to the embedder.

EmbedderReceiver
The port that the embedder receives messages on.

FilterPattern
Filter for file selection; the String content is expected to be extension (e.g, "doc", without the prefixing ".")

MediaMetadata
https://w3c.github.io/mediasession/#mediametadata

MediaPositionState
https://w3c.github.io/mediasession/#dictdef-mediapositionstate

Enums

ContextMenuResult

Cursor
A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

EmbedderMsg

MediaSessionEvent
Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState
https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

PermissionName
Enum with variants that match the DOM PermissionName enum

PermissionPrompt
Information required to display a permission prompt

PermissionRequest
Status for prompting user for permission.

PromptDefinition

PromptOrigin

PromptResult

Constants

_IMPL_NUM_FromPrimitive_FOR_Cursor

Traits

EventLoopWaker