Crate embedder_traits

Modules

resources

Structs

DualRumbleEffectParams

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype-dual-rumble

EmbedderProxy

Sends messages to the embedder.

EmbedderReceiver

The port that the embedder receives messages on.

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

MediaMetadata

https://w3c.github.io/mediasession/#mediametadata

MediaPositionState

https://w3c.github.io/mediasession/#dictdef-mediapositionstate

Enums

CompositorEventVariant

The variant of CompositorEvent that was delivered to a pipeline.

ContextMenuResult

Cursor

A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

EmbedderMsg

GamepadHapticEffectType

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype

InputMethodType

Used to specify the kind of input method editor appropriate to edit a field. This is a subset of htmlinputelement::InputType because some variants of InputType don’t make sense in this context.

MediaSessionEvent

Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState

https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

PermissionName

Enum with variants that match the DOM PermissionName enum

PermissionPrompt

Information required to display a permission prompt

PermissionRequest

Status for prompting user for permission.

PromptDefinition

PromptOrigin

PromptResult

Traits

EventLoopWaker