embedder_traits

Module input_events

Structs

GamepadIndex

Index of gamepad in list of system’s connected gamepads

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad

InputEventAndId

InputEventId

InputEventResult

KeyboardEvent

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

ScrollEvent

TouchEvent

TouchId

An opaque identifier for a touch point.

WheelDelta

The Wheel event deltas in every direction

WheelEvent

Enums

EditingActionEvent

An editing action that should be performed on a WebView.

GamepadEvent

The type of Gamepad event

GamepadUpdateType

The type of Gamepad input being updated

ImeEvent

InputEvent

An input event that is sent from the embedder to Servo.

MouseButton

MouseButtonAction

The types of mouse events

TouchEventType

The type of input represented by a multi-touch event.

WheelMode

Mode to measure WheelDelta floats in

Statics

INPUT_EVENT_ID 🔒